start test framework
